fixed email and docker credentials in jenkins file

+++ Jenkinsfile
@@ -1,10 +1,6 @@
 pipeline {
     agent any
-    triggers {
-            //this will trigger this jenkins job when the merger happens
-            githubPush()
-            
-        }
+
     environment { 
         DOCKER_CREDENTIALS_ID = 'dockerhub-creds'
         DOCKER_IMAGE = "faizan262/mlops_assignment1:latest"
@@ -41,10 +37,20 @@ pipeline {
         stage('Deploy Container') {
             steps {
                 script {
+                    
+                    def containerExist = bat(script: 'docker ps -aq -f name=flask-app', returnStdout: true).trim()
+
+                    if (containerExist) {
+                        
+                        bat '''
+                            docker stop flask-app || exit /b 0
+                            docker rm flask-app || exit /b 0
+                        '''
+                    }
+
+                
                     bat '''
-                    docker stop flask-app || exit /b 0
-                    docker rm flask-app || exit /b 0
-                    docker run -d -p 5000:5000 --name flask-app %DOCKER_IMAGE%
+                        docker run -d -p 5003:5003 --name flask-app %DOCKER_IMAGE%
                     '''
                 }
             }
@@ -56,19 +62,19 @@ pipeline {
             emailext (
                 subject: 'Successful Deployment',
                 body: '''
-                    Flask ML app has been successfully deployed via Jenkins!
+                    Flask app has been successfully deployed using Jenkins!
 
                     Image: ${DOCKER_IMAGE}
                     Container Name: flask-app
                 ''',
-                to: 'faizan.official262@gmai.com'
+                to: 'faizan.official262@gmail.com'
             )
         }
         failure {
             emailext (
                 subject: 'Failed Deployment',
                 body: 'Check Jenkins logs for more details due to failure of email being not sent',
-                to: 'faizan.official262@gmai.com'
+                to: 'faizan.official262@gmail.com'
             )
         }
     }
